add a module check and a nicer error message when trying to build prod
